refactor: simplify publicConfig system
