+++ app/entry.client.tsx
@@ -1,11 +1,11 @@
 import { RemixBrowser } from "@remix-run/react";
 import { hydrateRoot } from "react-dom/client";
 import { registerServiceWorker } from "./misc/register-service-worker.client";
-import { initializeConfigClient } from "./utils/config";
+import { initializePublicConfigClient } from "./utils/config-public";
 
 function main() {
   registerServiceWorker();
-  initializeConfigClient();
+  initializePublicConfigClient();
   hydrateRoot(window.document, <RemixBrowser />);
 }
