+++ app/entry.server.tsx
@@ -2,7 +2,6 @@ import { RemixServer } from "@remix-run/react";
 import type { HandleDocumentRequestFunction } from "@remix-run/server-runtime";
 import { renderToString } from "react-dom/server";
 import { injectInitializeServer } from "./misc/initialize-server";
-import { injectConfigScript } from "./utils/config";
 
 injectInitializeServer();
 
@@ -16,7 +15,6 @@ const handler: HandleDocumentRequestFunction = (
   let markup = renderToString(
     <RemixServer context={remixContext} url={request.url} />
   );
-  markup = injectConfigScript(markup);
   responseHeaders.set("content-type", "text/html");
   return new Response("<!DOCTYPE html>" + markup, {
     status: responseStatusCode,
